# server/docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateWalletRequest:
    properties:
      userId:
        type: string
    type: object
  handler.CreateWalletResponse:
    properties:
      address:
        type: string
      id:
        type: string
      userId:
        type: string
    type: object
  handler.ExecuteTradeRequest:
    properties:
      amountIn:
        description: BNB for BUY, Token for SELL
        type: string
      amountOut:
        type: string
      decisionReason:
        type: string
      force:
        type: boolean
      goldenDogScore:
        type: integer
      profitLoss:
        type: number
      strategyUsed:
        type: string
      tokenAddress:
        type: string
      tokenSymbol:
        type: string
      type:
        description: BUY | SELL
        type: string
      userId:
        type: string
    type: object
...
  handler.WalletBalanceResponse:
    properties:
      address:
        type: string
      balance:
        type: number
      userId:
        type: string
    type: object
  handler.WalletConfigRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      userId:
        type: string
    type: object
  handler.WithdrawRequest:
    properties:
      amount:
        type: number
      userId:
        type: string
    type: object
  handler.WithdrawResponse:
    properties:
      address:
        type: string
      balance:
        type: number
      userId:
        type: string
    type: object
...
info:
  contact: {}
  description: EasyMeme server API for token analysis and discovery.
  title: EasyMeme API
  version: "0.1"
paths:
...
  /api/wallet/balance:
    get:
      description: Get managed wallet balance by user
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.WalletBalanceResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get managed wallet balance
      tags:
      - wallet
  /api/wallet/config:
    post:
      description: Upsert auto-trade config for managed wallet
      parameters:
      - description: Wallet config payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.WalletConfigRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upsert wallet config
      tags:
      - wallet
  /api/wallet/create:
    post:
      description: Create a managed wallet and store encrypted private key
      parameters:
      - description: Create wallet payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.CreateWalletRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.CreateWalletResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create managed wallet
      tags:
      - wallet
  /api/wallet/execute-trade:
    post:
      description: Execute managed wallet trade on BSC
      parameters:
      - description: Execute trade payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.ExecuteTradeRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Execute trade
      tags:
      - wallet
  /api/wallet/info:
    get:
      description: Get managed wallet address and balance (userId optional, fallback
        to EASYMEME_USER_ID)
      parameters:
      - description: User ID
        in: query
        name: userId
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.WalletBalanceResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get managed wallet info
      tags:
      - wallet
  /api/wallet/withdraw:
    post:
      description: Decrease managed wallet balance (placeholder, no on-chain tx)
      parameters:
      - description: Withdraw payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.WithdrawRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.WithdrawResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Withdraw from managed wallet
      tags:
      - wallet
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"